+++ matcher/src/kafka/KafkaOrderBookProducer.cpp
@@ -18,8 +18,9 @@ namespace kafka
         long long ms = std::chrono::duration_cast<std::chrono::milliseconds>(s.timestamp().time_since_epoch()).count();
         // Prepare bids and asks as arrays of {price, volume}
         nlohmann::json bids = nlohmann::json::array();
-        std::map<double, double, std::greater<>> bid_vols = s.bidVolumes();
-        for (std::map<double, double, std::greater<>>::const_iterator it = bid_vols.begin(); it != bid_vols.end(); ++it)
+        // Use int for volume values
+        std::map<double, int, std::greater<>> bid_vols = s.bidVolumes();
+        for (std::map<double, int, std::greater<>>::const_iterator it = bid_vols.begin(); it != bid_vols.end(); ++it)
         {
             nlohmann::json entry;
             entry["price"] = it->first;
@@ -27,8 +28,8 @@ namespace kafka
             bids.push_back(entry);
         }
         nlohmann::json asks = nlohmann::json::array();
-        std::map<double, double> ask_vols = s.askVolumes();
-        for (std::map<double, double>::const_iterator it = ask_vols.begin(); it != ask_vols.end(); ++it)
+        std::map<double, int> ask_vols = s.askVolumes();
+        for (std::map<double, int>::const_iterator it = ask_vols.begin(); it != ask_vols.end(); ++it)
         {
             nlohmann::json entry;
             entry["price"] = it->first;
